+++ libyul/AsmAnalysis.cpp
@@ -75,7 +75,7 @@ AsmAnalysisInfo AsmAnalyzer::analyzeStrictAssertCorrect(Dialect const& _dialect,
 		{},
 		_object.dataNames()
 	).analyze(*_object.code);
-	yulAssert(success && errorList.empty(), "Invalid assembly/yul code.");
+	yulAssert(success && !errors.hasErrors(), "Invalid assembly/yul code.");
 	return analysisInfo;
 }
 
@@ -263,6 +263,8 @@ vector<YulString> AsmAnalyzer::operator()(FunctionCall const& _funCall)
 		returnTypes = &f->returns;
 		if (f->literalArguments)
 			needsLiteralArguments = &f->literalArguments.value();
+
+		warnOnInstructions(_funCall.functionName.name.str(), _funCall.functionName.location);
 	}
 	else if (!m_currentScope->lookup(_funCall.functionName.name, GenericVisitor{
 		[&](Scope::Variable const&)
@@ -571,6 +573,16 @@ bool AsmAnalyzer::warnOnInstructions(evmasm::Instruction _instr, SourceLocation 
 	else if (_instr == evmasm::Instruction::CHAINID && !m_evmVersion.hasChainID())
 	{
 		errorForVM("only available for Istanbul-compatible");
+	}
+	else if (_instr == evmasm::Instruction::PC)
+	{
+		m_errorReporter.warning(
+			2450_error,
+			_location,
+			"The \"" +
+			boost::to_lower_copy(instructionInfo(_instr).name) +
+			"\" instruction is deprecated and will be removed in the next breaking release."
+		);
 	}
 	else if (_instr == evmasm::Instruction::SELFBALANCE && !m_evmVersion.hasSelfBalance())
 	{